requirements.txt
discord.py
dotenv
openai